# docs/swag/wallet/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Wallet API server",
        "title": "Wallet API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ap11.0dev.cc",
    "basePath": "/",
    "paths": {
...
        "/bo/refund/manual_process": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "manual refund processed by admin when auto refund is failed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ManualRefund"
                ],
                "summary": "Process manual refund processed by admin",
                "parameters": [
                    {
                        "description": "Mnual refund request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.ManualRefundArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully refund by admin",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Request invalid",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Wallet user not found",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/envelope": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new envelope (red packet) that can be claimed by users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Envelope"
                ],
                "summary": "Create envelope",
                "parameters": [
                    {
                        "description": "Envelope creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created envelope",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid envelope data or insufficient balance",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/envelope/autoRefund": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Process automatic refund for all expired envelopes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Envelope"
                ],
                "summary": "Auto refund envelopes",
                "responses": {
                    "200": {
                        "description": "Successfully processed auto refund",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/envelope/claim": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Claim an available envelope (red packet) to receive money",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Envelope"
                ],
                "summary": "Claim envelope",
                "parameters": [
                    {
                        "description": "Envelope claim request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeClaimRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully claimed envelope",
                        "schema": {
                            "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeClaimResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid envelope ID or wallet ID, or envelope already claimed",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Envelope not found",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/envelope/refund/:envelopeID": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Process refund for all expired envelopeByID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Envelope"
                ],
                "summary": "refund envelopeByID",
                "responses": {
                    "200": {
                        "description": "Successfully processed auto refund",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeClaimRequest": {
            "type": "object",
            "required": [
                "envelopeId"
            ],
            "properties": {
                "envelopeId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                },
                "walletId": {
                    "type": "integer"
                }
            }
        },
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeClaimResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "claimed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "envelope": {
                    "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_model.Envelope"
                },
                "envelope_detail_id": {
                    "type": "integer"
                },
                "envelope_detail_status": {
                    "type": "string"
                },
                "envelope_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wallet": {
                    "$ref": "#/definitions/bitbucket_org_interdigital_aka-im-server-wallet_internal_model.Wallet"
                }
            }
        },
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeCreateRequest": {
            "type": "object",
            "required": [
                "envelopeType",
                "remarks",
                "totalAmount",
                "totalClaimer",
                "walletId"
            ],
            "properties": {
                "envelopeType": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "toUserId": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                },
                "totalClaimer": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                },
                "walletId": {
                    "type": "integer"
                }
            }
        },
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.EnvelopeCreateResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "deletedBy": {
                    "type": "string"
                },
                "envelopeID": {
                    "type": "integer"
                },
                "envelopeType": {
                    "type": "string"
                },
                "expiredAt": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxNumReceived": {
                    "type": "integer"
                },
                "refundedAt": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                },
                "totalAmountClaimed": {
                    "type": "number"
                },
                "totalAmountRefunded": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "walletID": {
                    "type": "integer"
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_domain.ManualRefundArgs": {
            "type": "object",
            "properties": {
                "envelopeIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "transferIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_model.Envelope": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "deleted_by": {
                    "type": "string"
                },
                "envelopeType": {
                    "type": "string"
                },
                "envelope_id": {
                    "type": "integer"
                },
                "expired_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_num_received": {
                    "type": "integer"
                },
                "refunded_at": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_amount_claimed": {
                    "type": "number"
                },
                "total_amount_refunded": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        },
        "bitbucket_org_interdigital_aka-im-server-wallet_internal_model.Wallet": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "deleted_by": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}